web-api/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
grpcio==1.31.0
grpcio-tools==1.31.0
protobuf==3.12.4
six==1.15.0